Created warnings field in user schema. Updated unit tests to fix.
