+++ backend/app/data/users.json
@@ -5,7 +5,8 @@
     "hashed_password": "$2b$12$nCqzSBYg3ojbbx8WZXENmeSZVdFmrisWpWwj/nRSGBjfjTiachis.",
     "role": "admin",
     "created_at": "2025-10-20T23:09:02.346841",
-    "active": true
+    "active": true,
+    "warnings": 0
   },
   {
     "id": "77d60efe-0870-4323-b164-3d76474f251f",
@@ -13,6 +14,7 @@
     "hashed_password": "$2b$12$B6mJk/6rXrgrTFnq9YseZu6AIvuEZr1O2lKI6kIa2BJyjyHWz6Kya",
     "role": "user",
     "created_at": "2025-10-23T17:10:27.600472",
-    "active": true
+    "active": true,
+    "warnings": 0
   }
 ]
