+++ backend/tests/test_user_schema.py
@@ -1,19 +1,6 @@
 import pytest
 from pydantic import ValidationError
 from app.schemas.user import User, UserCreate, UserUpdate
-import datetime
-
-@pytest.fixture
-def user_data():
-    payload = {
-        "id": "1234",
-        "username": "testmovielover",
-        "hashed_password": "$2y$10$b1d2DgDhd1bdRbiwSqfZs.MhtyNCMHaYbQp3.6D3ngYLQ9ySTM/HO",
-        "role": "user",
-        "created_at": datetime.datetime.now(),
-        "active": True
-    }
-    return payload
 
 def test_missing_required_fields():
     with pytest.raises(ValidationError):
@@ -23,6 +10,10 @@ def test_valid_user(user_data):
     user = User(**user_data)
     assert user.username == user_data["username"]
     assert user.active is True
+
+def test_valid_user_default_warnings(user_data):
+    user = User(**user_data)
+    assert user.warnings == user_data["warnings"]
 
 def test_username_too_short(user_data):
     with pytest.raises(ValidationError):
